fix(trace): extract shared session scanner and fix trace list NUT cwd
Move the ended-session filesystem scan into src/agentSessionScanner.ts
so all three trace commands (list, read, delete) share the same logic
and no longer miss sessions that were ended via agent preview end.
Also add missing cwd to the z4.agent.trace.list NUT before() hook.
